requirements.txt
asgiref==3.7.2
backports.zoneinfo;python_version<"3.9"
blessed==1.20.0
bpython==0.24
certifi==2024.2.2
charset-normalizer==3.3.2
curtsies==0.4.2
cwcwidth==0.1.9
Django==4.2.9
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
Markdown==3.5.2
psycopg2-binary==2.9.9
pygments==2.17.2
PyJWT==2.8.0
pytz==2024.1
pyxdg==0.28
requests==2.31.0
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
urllib3==2.2.0
wcwidth==0.2.13
zipp==3.17.0
